+++ govcd/media_test.go
@@ -7,6 +7,7 @@
 package govcd
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 
@@ -15,6 +16,8 @@ import (
 
 // Tests System function UploadMediaImage by checking if provided standard iso file uploaded.
 func (vcd *TestVCD) Test_UploadMediaImage(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
 	skipWhenMediaPathMissing(vcd, check)
 
 	uploadTask, err := vcd.vdc.UploadMediaImage(TestUploadMedia, "upload from test", vcd.config.Media.MediaPath, 1024)
@@ -34,7 +37,7 @@ func skipWhenMediaPathMissing(vcd *TestVCD, check *C) {
 }
 
 func verifyMediaImageUploaded(vdc *Vdc, check *C, itemName string) {
-	results, err := queryMediaItemsWithFilter(vdc, "name=="+itemName)
+	results, err := queryMediaWithFilter(vdc, "name=="+itemName)
 
 	check.Assert(err, Equals, nil)
 	check.Assert(len(results), Equals, 1)
@@ -42,6 +45,8 @@ func verifyMediaImageUploaded(vdc *Vdc, check *C, itemName string) {
 
 // Tests System function UploadMediaImage by checking UploadTask.GetUploadProgress returns values of progress.
 func (vcd *TestVCD) Test_UploadMediaImage_progress_works(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
 	skipWhenMediaPathMissing(vcd, check)
 	itemName := TestUploadMedia + "2"
 
@@ -64,6 +69,8 @@ func (vcd *TestVCD) Test_UploadMediaImage_progress_works(check *C) {
 
 // Tests System function UploadMediaImage by checking UploadTask.ShowUploadProgress writes values of progress to stdin.
 func (vcd *TestVCD) Test_UploadMediaImage_ShowUploadProgress_works(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
 	skipWhenMediaPathMissing(vcd, check)
 	itemName := TestUploadMedia + "3"
 
@@ -95,6 +102,8 @@ func (vcd *TestVCD) Test_UploadMediaImage_ShowUploadProgress_works(check *C) {
 // Tests System function UploadMediaImage by creating media item and expecting specific error
 // then trying to create same media item. As vCD returns cryptic error for such case.
 func (vcd *TestVCD) Test_UploadMediaImage_error_withSameItem(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
 	skipWhenMediaPathMissing(vcd, check)
 	itemName := TestUploadMedia + "4"
 
@@ -112,38 +121,152 @@ func (vcd *TestVCD) Test_UploadMediaImage_error_withSameItem(check *C) {
 
 // Tests System function Delete by creating media item and
 // deleting it after.
-func (vcd *TestVCD) Test_DeleteMediaImage(check *C) {
-	skipWhenMediaPathMissing(vcd, check)
-	itemName := TestUploadMedia + "5"
+func (vcd *TestVCD) Test_DeleteMedia(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
 
-	uploadTask, err := vcd.vdc.UploadMediaImage(itemName, "upload from test", vcd.config.Media.MediaPath, 1024)
+	skipWhenMediaPathMissing(vcd, check)
+	if vcd.config.VCD.Org == "" {
+		check.Skip("Test_DeleteMedia: Org name not given")
+		return
+	}
+	if vcd.config.VCD.Catalog.Name == "" {
+		check.Skip("Test_DeleteMedia: Catalog name not given")
+		return
+	}
+	org, err := vcd.client.GetOrgByName(vcd.config.VCD.Org)
+	check.Assert(err, IsNil)
+	check.Assert(org, NotNil)
+
+	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
+	check.Assert(err, IsNil)
+	check.Assert(catalog, NotNil)
+
+	itemName := "TestDeleteMedia"
+	uploadTask, err := catalog.UploadMediaImage(itemName, "upload from test", vcd.config.Media.MediaPath, 1024)
 	check.Assert(err, IsNil)
 	err = uploadTask.WaitTaskCompletion()
 	check.Assert(err, IsNil)
 
-	AddToCleanupList(itemName, "mediaImage", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name, "Test_UploadMediaImage")
+	AddToCleanupList(itemName, "mediaCatalogImage", vcd.org.Org.Name+"|"+vcd.config.VCD.Catalog.Name, "Test_DeleteMediaImage")
 
-	mediaItem, err := vcd.vdc.FindMediaImage(itemName)
+	media, err := catalog.GetMediaByName(itemName, true)
 	check.Assert(err, IsNil)
-	check.Assert(mediaItem, Not(Equals), MediaItem{})
+	check.Assert(media, NotNil)
+	check.Assert(media.Media.Name, Equals, itemName)
 
-	task, err := mediaItem.Delete()
+	task, err := media.Delete()
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
 
-	mediaItem, err = vcd.vdc.FindMediaImage(itemName)
-	check.Assert(err, IsNil)
-	check.Assert(mediaItem, Equals, MediaItem{})
-
+	_, err = catalog.GetMediaByName(itemName, true)
+	check.Assert(err, NotNil)
+	check.Assert(IsNotFound(err), Equals, true)
 }
 
-// Tests System function FindMediaAsCatalogItem by creating media item and
-// and finding it as catalog item after.
-func (vcd *TestVCD) Test_FindMediaAsCatalogItem(check *C) {
-	skipWhenMediaPathMissing(vcd, check)
-	itemName := TestUploadMedia + "6"
+// Tests System function Refresh
+func (vcd *TestVCD) Test_RefreshMedia(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
 
+	skipWhenMediaPathMissing(vcd, check)
+	if vcd.config.VCD.Org == "" {
+		check.Skip("Test_RefreshMedia: Org name not given")
+		return
+	}
+	if vcd.config.VCD.Catalog.Name == "" {
+		check.Skip("Test_RefreshMedia: Catalog name not given")
+		return
+	}
+	if vcd.config.Media.Media == "" {
+		check.Skip("Test_RefreshMediaRecord: Media name not given")
+		return
+	}
+	org, err := vcd.client.GetOrgByName(vcd.config.VCD.Org)
+	check.Assert(err, IsNil)
+	check.Assert(org, NotNil)
+
+	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
+	check.Assert(err, IsNil)
+	check.Assert(catalog, NotNil)
+
+	itemName := vcd.config.Media.Media
+
+	media, err := catalog.GetMediaByName(itemName, true)
+	check.Assert(err, IsNil)
+	check.Assert(media, NotNil)
+	check.Assert(media.Media.Name, Equals, itemName)
+
+	oldMedia := media
+	err = media.Refresh()
+	check.Assert(err, IsNil)
+
+	check.Assert(media, NotNil)
+	check.Assert(oldMedia.Media.ID, Equals, media.Media.ID)
+	check.Assert(oldMedia.Media.Name, Equals, media.Media.Name)
+	check.Assert(oldMedia.Media.HREF, Equals, media.Media.HREF)
+}
+
+// Tests Media retrieval by name, by ID, and by a combination of name and ID
+func (vcd *TestVCD) Test_GetMedia(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
+	if vcd.config.VCD.Org == "" {
+		check.Skip("Test_GetMedia: Org name not given")
+		return
+	}
+	if vcd.config.VCD.Catalog.Name == "" {
+		check.Skip("Test_GetMedia: Catalog name not given")
+		return
+	}
+	org, err := vcd.client.GetOrgByName(vcd.config.VCD.Org)
+	check.Assert(err, IsNil)
+	check.Assert(org, NotNil)
+
+	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
+	check.Assert(err, IsNil)
+	check.Assert(catalog, NotNil)
+
+	getByName := func(name string, refresh bool) (genericEntity, error) {
+		err = catalog.Refresh()
+		check.Assert(err, IsNil)
+		return catalog.GetMediaByName(name, refresh)
+	}
+	getById := func(id string, refresh bool) (genericEntity, error) {
+		return catalog.GetMediaById(id)
+	}
+	getByNameOrId := func(id string, refresh bool) (genericEntity, error) {
+		return catalog.GetMediaByNameOrId(id, refresh)
+	}
+
+	var def = getterTestDefinition{
+		parentType:    "Catalog",
+		parentName:    vcd.config.VCD.Catalog.Name,
+		entityType:    "Media",
+		entityName:    vcd.config.Media.Media,
+		getByName:     getByName,
+		getById:       getById,
+		getByNameOrId: getByNameOrId,
+	}
+	vcd.testFinderGetGenericEntity(def, check)
+}
+
+func (vcd *TestVCD) Test_QueryMedia(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
+	skipWhenMediaPathMissing(vcd, check)
+
+	if vcd.config.VCD.Org == "" {
+		check.Skip("Test_QueryMedia: Org name not given")
+		return
+	}
+	if vcd.config.VCD.Catalog.Name == "" {
+		check.Skip("Test_QueryMedia: Catalog name not given")
+		return
+	}
+	if vcd.config.Media.Media == "" {
+		check.Skip("Test_RefreshMediaRecord: Media name not given")
+		return
+	}
 	// Fetching organization
 	org, err := vcd.client.GetAdminOrgByName(vcd.org.Org.Name)
 	check.Assert(err, IsNil)
@@ -152,47 +275,61 @@ func (vcd *TestVCD) Test_FindMediaAsCatalogItem(check *C) {
 	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
 	check.Assert(err, IsNil)
 
-	uploadTask, err := catalog.UploadMediaImage(itemName, "upload from test", vcd.config.Media.MediaPath, 1024)
-	check.Assert(err, IsNil)
-	err = uploadTask.WaitTaskCompletion()
-	check.Assert(err, IsNil)
+	testQueryMediaName := vcd.config.Media.Media
 
-	AddToCleanupList(itemName, "mediaImage", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name, "Test_UploadMediaImage")
-
-	err = vcd.org.Refresh()
+	media, err := catalog.QueryMedia(testQueryMediaName)
 	check.Assert(err, IsNil)
+	check.Assert(media, Not(Equals), nil)
 
-	catalogItem, err := FindMediaAsCatalogItem(vcd.org, vcd.config.VCD.Catalog.Name, itemName)
-	check.Assert(err, IsNil)
-	check.Assert(catalogItem, Not(Equals), CatalogItem{})
-	check.Assert(catalogItem.CatalogItem.Name, Equals, itemName)
+	check.Assert(media.MediaRecord.Name, Equals, testQueryMediaName)
+	check.Assert(media.MediaRecord.HREF, Not(Equals), "")
 
+	// find Invalid media
+	media, err = catalog.QueryMedia("INVALID")
+	check.Assert(IsNotFound(err), Equals, true)
+	check.Assert(media, IsNil)
 }
 
 // Tests System function Refresh
-func (vcd *TestVCD) Test_RefreshMediaImage(check *C) {
+func (vcd *TestVCD) Test_RefreshMediaRecord(check *C) {
+	fmt.Printf("Running: %s\n", check.TestName())
+
 	skipWhenMediaPathMissing(vcd, check)
-	itemName := "TestRefreshMedia"
+	if vcd.config.VCD.Org == "" {
+		check.Skip("Test_RefreshMediaRecord: Org name not given")
+		return
+	}
+	if vcd.config.VCD.Catalog.Name == "" {
+		check.Skip("Test_RefreshMediaRecord: Catalog name not given")
+		return
+	}
 
-	uploadTask, err := vcd.vdc.UploadMediaImage(itemName, "upload from test", vcd.config.Media.MediaPath, 1024)
+	if vcd.config.Media.Media == "" {
+		check.Skip("Test_RefreshMediaRecord: Media name not given")
+		return
+	}
+
+	org, err := vcd.client.GetOrgByName(vcd.config.VCD.Org)
 	check.Assert(err, IsNil)
-	check.Assert(uploadTask, NotNil)
-	err = uploadTask.WaitTaskCompletion()
+	check.Assert(org, NotNil)
+
+	catalog, err := org.GetCatalogByName(vcd.config.VCD.Catalog.Name, false)
+	check.Assert(err, IsNil)
+	check.Assert(catalog, NotNil)
+
+	itemName := vcd.config.Media.Media
+
+	mediaRecord, err := catalog.QueryMedia(itemName)
+	check.Assert(err, IsNil)
+	check.Assert(mediaRecord, NotNil)
+	check.Assert(mediaRecord.MediaRecord.Name, Equals, itemName)
+
+	oldMediaRecord := mediaRecord
+	err = mediaRecord.Refresh()
 	check.Assert(err, IsNil)
 
-	AddToCleanupList(itemName, "mediaImage", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name, "Test_UploadMediaImage")
-
-	mediaItem, err := vcd.vdc.FindMediaImage(itemName)
-	check.Assert(err, IsNil)
-	check.Assert(mediaItem, NotNil)
-	check.Assert(mediaItem, Not(Equals), MediaItem{})
-
-	oldMediaItem := mediaItem
-	err = mediaItem.Refresh()
-	check.Assert(err, IsNil)
-
-	check.Assert(mediaItem, NotNil)
-	check.Assert(oldMediaItem.MediaItem.ID, Equals, mediaItem.MediaItem.ID)
-	check.Assert(oldMediaItem.MediaItem.Name, Equals, mediaItem.MediaItem.Name)
-	check.Assert(oldMediaItem.MediaItem.HREF, Equals, mediaItem.MediaItem.HREF)
+	check.Assert(mediaRecord, NotNil)
+	check.Assert(oldMediaRecord.MediaRecord.ID, Equals, mediaRecord.MediaRecord.ID)
+	check.Assert(oldMediaRecord.MediaRecord.Name, Equals, mediaRecord.MediaRecord.Name)
+	check.Assert(oldMediaRecord.MediaRecord.HREF, Equals, mediaRecord.MediaRecord.HREF)
 }